requirements.txt
Flask
tensorflow==2.11.0
pillow
numpy
gunicorn
